+++ node/src/tasks/bsp_volunteer_mock.rs
@@ -1,13 +1,7 @@
-use anyhow::anyhow;
-use log::{debug, error, info};
-use shc_common::types::{FileKey, FileMetadata, HasherOutT};
-use sp_trie::TrieLayout;
 use std::str::FromStr;
 
-use sc_network::PeerId;
+use log::{debug, error, info};
 use sp_core::H256;
-
-use file_manager::traits::FileStorage;
 use storage_hub_infra::{actor::ActorHandle, event_bus::EventHandler};
 
 use crate::services::{
@@ -15,7 +9,6 @@ use crate::services::{
         commands::BlockchainServiceInterface, events::NewStorageRequest,
         handler::BlockchainService, types::ExtrinsicResult,
     },
-    file_transfer::commands::FileTransferServiceInterface,
     handler::{StorageHubHandler, StorageHubHandlerConfig},
 };
 
@@ -23,14 +16,12 @@ const LOG_TARGET: &str = "bsp-volunteer-mock-task";
 
 pub struct BspVolunteerMockTask<SHC: StorageHubHandlerConfig> {
     storage_hub_handler: StorageHubHandler<SHC>,
-    file_key_cleanup: Option<HasherOutT<SHC::TrieLayout>>,
 }
 
 impl<SHC: StorageHubHandlerConfig> Clone for BspVolunteerMockTask<SHC> {
     fn clone(&self) -> BspVolunteerMockTask<SHC> {
         Self {
             storage_hub_handler: self.storage_hub_handler.clone(),
-            file_key_cleanup: self.file_key_cleanup.clone(),
         }
     }
 }
@@ -39,7 +30,6 @@ impl<SHC: StorageHubHandlerConfig> BspVolunteerMockTask<SHC> {
     pub fn new(storage_hub_handler: StorageHubHandler<SHC>) -> Self {
         Self {
             storage_hub_handler,
-            file_key_cleanup: None,
         }
     }
 }
@@ -53,24 +43,6 @@ impl<SHC: StorageHubHandlerConfig> EventHandler<NewStorageRequest> for BspVolunt
             event.fingerprint
         );
 
-        let result = self.inner_handle_event(event).await;
-        if result.is_err() {
-            if let Some(file_key) = &self.file_key_cleanup {
-                let _ = self
-                    .storage_hub_handler
-                    .file_transfer
-                    .unregister_file(file_key.as_ref().into())
-                    .await;
-                let mut write_file_storage = self.storage_hub_handler.file_storage.write().await;
-                write_file_storage.delete_file(file_key);
-            }
-        }
-        result
-    }
-}
-
-impl<SHC: StorageHubHandlerConfig> BspVolunteerMockTask<SHC> {
-    async fn inner_handle_event(&self, event: NewStorageRequest) -> anyhow::Result<()> {
         let fingerprint: [u8; 32] = event
             .fingerprint
             .as_ref()
@@ -80,7 +52,7 @@ impl<SHC: StorageHubHandlerConfig> BspVolunteerMockTask<SHC> {
         // Build extrinsic.
         let call =
             storage_hub_runtime::RuntimeCall::FileSystem(pallet_file_system::Call::bsp_volunteer {
-                location: event.location.clone(),
+                location: event.location,
                 fingerprint: fingerprint.into(),
             });
 
@@ -90,42 +62,13 @@ impl<SHC: StorageHubHandlerConfig> BspVolunteerMockTask<SHC> {
             .send_extrinsic(call)
             .await?;
 
-        // Construct file metadata.
-        let metadata = FileMetadata {
-            owner: event.who.to_string(),
-            size: event.size as u64,
-            fingerprint: event.fingerprint,
-            location: event.location.to_vec(),
-        };
-
-        // Get the file key.
-        let file_key: FileKey = metadata
-            .key::<<SHC::TrieLayout as TrieLayout>::Hash>()
-            .as_ref()
-            .try_into()?;
-
-        // Optimistically register the file for upload in the file transfer service.
-        // This solves the race condition between the user and the BSP, where the user could react faster
-        // to the BSP volunteering than the BSP, and therefore initiate a new upload request before the
-        // BSP has registered the file and peer ID in the file transfer service.
-        for peer_id in event.user_peer_ids.iter() {
-            let peer_id = PeerId::from_bytes(peer_id.as_slice())
-                .map_err(|_| anyhow!("PeerId should be valid; qed"))?;
-            self.storage_hub_handler
-                .file_transfer
-                .register_new_file_peer(peer_id.clone(), file_key.clone())
-                .await
-                .map_err(|_| anyhow!("Failed to register peer file."))?;
-        }
-
         // Wait for the transaction to be included in a block.
         let mut block_hash = None;
         // TODO: Consider adding a timeout.
         while let Some(tx_result) = tx_watcher.recv().await {
             // Parse the JSONRPC string, now that we know it is not an error.
-            let json: serde_json::Value = serde_json::from_str(&tx_result).map_err(|_| {
-                anyhow!("The result, if not an error, can only be a JSONRPC string; qed")
-            })?;
+            let json: serde_json::Value = serde_json::from_str(&tx_result)
+                .expect("The result, if not an error, can only be a JSONRPC string; qed");
 
             debug!(target: LOG_TARGET, "Transaction information: {:?}", json);
 
@@ -136,7 +79,7 @@ impl<SHC: StorageHubHandlerConfig> BspVolunteerMockTask<SHC> {
                 block_hash = Some(H256::from_str(in_block)?);
                 let subscription_id = json["params"]["subscription"]
                     .as_number()
-                    .ok_or_else(|| anyhow!("Subscription should exist and be a number; qed"))?;
+                    .expect("Subscription should exist and be a number; qed");
 
                 // Unwatch extrinsic to release tx_watcher.
                 self.storage_hub_handler
@@ -153,9 +96,9 @@ impl<SHC: StorageHubHandlerConfig> BspVolunteerMockTask<SHC> {
         }
 
         // Get the extrinsic from the block, with its events.
-        let block_hash = block_hash.ok_or_else(
-            || anyhow!("Block hash should exist after waiting for extrinsic to be included in a block; qed")
-        )?;
+        let block_hash = block_hash.expect(
+            "Block hash should exist after waiting for extrinsic to be included in a block; qed",
+        );
         let extrinsic_in_block = self
             .storage_hub_handler
             .blockchain
@@ -165,7 +108,7 @@ impl<SHC: StorageHubHandlerConfig> BspVolunteerMockTask<SHC> {
         // Check if the extrinsic was successful. In this mocked task we know this should fail if Alice is
         // not a registered BSP.
         let extrinsic_successful = ActorHandle::<BlockchainService>::extrinsic_result(extrinsic_in_block.clone())
-            .map_err(|_| anyhow!("Extrinsic does not contain an ExtrinsicFailed nor ExtrinsicSuccess event, which is not possible; qed"))?;
+            .expect("Extrinsic does not contain an ExtrinsicFailed nor ExtrinsicSuccess event, which is not possible; qed");
         match extrinsic_successful {
             ExtrinsicResult::Success { dispatch_info } => {
                 info!(target: LOG_TARGET, "Extrinsic successful with dispatch info: {:?}", dispatch_info);
